proc: move cluster assembly to a new file
